updated a bunch of files for better readability
