+++ tbc/spells.go
@@ -1,5 +1,7 @@
 package tbc
 
+import "math"
+
 type Cast struct {
 	Spell *Spell
 	// Caster ... // Needed for onstruck effects?
@@ -24,6 +26,8 @@ type Cast struct {
 	Effects []AuraEffect // effects applied ONLY to this cast.
 }
 
+// NewCast constructs a Cast from the current simulation and selected spell.
+//  OnCast mechanics are applied at this time (anything that modifies the cast before its cast, usually just mana cost stuff)
 func NewCast(sim *Simulation, sp *Spell) *Cast {
 	cast := &Cast{
 		Spell:      sp,
@@ -33,11 +37,10 @@ func NewCast(sim *Simulation, sp *Spell) *Cast {
 	}
 
 	castTime := sp.CastTime
-	isLB := sp.ID == MagicIDLB12
-	isCL := sp.ID == MagicIDCL6
+	itsElectric := sp.ID == MagicIDLB12 || sp.ID == MagicIDCL6
 
-	if isLB || isCL {
-		// Talent to reduce cast time.
+	if itsElectric {
+		// TODO: Add LightningMaster to talent list (this will never not be selected for an elemental shaman)
 		castTime -= 0.5 // Talent Lightning Mastery
 	}
 	castTime /= (1 + ((sim.Stats[StatHaste] + sim.Buffs[StatHaste]) / 1576)) // 15.76 rating grants 1% spell haste
@@ -47,7 +50,7 @@ func NewCast(sim *Simulation, sp *Spell) *Cast {
 	cast.CastTime = castTime
 	cast.TicksUntilCast = int(castTime*float64(TicksPerSecond)) + 1 // round up
 
-	if isLB || isCL {
+	if itsElectric {
 		cast.ManaCost *= 1 - (0.02 * float64(sim.Options.Talents.Convection))
 	}
 
@@ -61,6 +64,65 @@ func NewCast(sim *Simulation, sp *Spell) *Cast {
 	return cast
 }
 
+// ChooseSpell is a basic rotation spell selector. This is the default
+// spell selection logic if not using the 'ai optimizer' for selecting spells.
+func ChooseSpell(sim *Simulation, _ bool) int {
+	if sim.RotationIdx == -1 {
+		lowestWait := math.MaxInt32
+		wasMana := false
+		for i := 0; i < len(sim.SpellRotation); i++ {
+			sp := sim.SpellRotation[i]
+			so := sp.ID
+			cast := NewCast(sim, sp)
+			if sim.CDs[so] > 0 { // if
+				if sim.CDs[so] < lowestWait {
+					lowestWait = sim.CDs[so]
+				}
+				continue
+			}
+			if sim.CurrentMana >= cast.ManaCost {
+				sim.CastingSpell = cast
+				return cast.TicksUntilCast
+			}
+			manaRegenTicks := int(math.Ceil((cast.ManaCost - sim.CurrentMana) / sim.manaRegen()))
+			if manaRegenTicks < lowestWait {
+				lowestWait = manaRegenTicks
+				wasMana = true
+			}
+		}
+		if wasMana && sim.metrics.OOMAt == 0 { // loop only completes if no spell was found.
+			sim.metrics.OOMAt = sim.CurrentTick / TicksPerSecond
+			sim.metrics.DamageAtOOM = sim.metrics.TotalDamage
+		}
+		return lowestWait
+	}
+
+	sp := sim.SpellRotation[sim.RotationIdx]
+	so := sp.ID
+	cast := NewCast(sim, sp)
+	if sim.CDs[so] < 1 {
+		if sim.CurrentMana >= cast.ManaCost {
+			sim.CastingSpell = cast
+			sim.RotationIdx++
+			if sim.RotationIdx == len(sim.SpellRotation) {
+				sim.RotationIdx = 0
+			}
+			return cast.TicksUntilCast
+		} else {
+			if sim.Debug != nil {
+				sim.Debug("Current Mana %0.0f, Cast Cost: %0.0f\n", sim.CurrentMana, cast.ManaCost)
+			}
+			if sim.metrics.OOMAt == 0 {
+				sim.metrics.OOMAt = sim.CurrentTick / TicksPerSecond
+				sim.metrics.DamageAtOOM = sim.metrics.TotalDamage
+			}
+			return int(math.Ceil((cast.ManaCost - sim.CurrentMana) / sim.manaRegen()))
+		}
+	}
+	return sim.CDs[so]
+}
+
+// Spell represents a single castable spell. This is all the data needed to begin a cast.
 type Spell struct {
 	ID         int32
 	Name       string
@@ -76,6 +138,7 @@ type Spell struct {
 	DotDur float64
 }
 
+// DamageType is currently unused.
 type DamageType byte
 
 const (
@@ -84,14 +147,15 @@ const (
 	DamageTypeNature
 	DamageTypeFrost
 
-	// who cares
+	// Who cares about these fake damage types.
 	DamageTypeShadow
 	DamageTypeHoly
 	DamageTypeArcane
 )
 
-// spells
-// TODO: DRP == (spellrankavailbetobetrained+11)/70
+// All Spells
+// TODO: Downrank Penalty == (spellrankavailbetobetrained+11)/70
+//    Might not even be worth calculating because I don't think there is much case for ever downranking.
 var spells = []Spell{
 	// {ID: MagicIDLB4, Name: "LB4", Coeff: 0.795, CastTime: 2.5, MinDmg: 88, MaxDmg: 100, Mana: 50, DamageType: DamageTypeNature},
 	// {ID: MagicIDLB10, Name: "LB10", Coeff: 0.795, CastTime: 2.5, MinDmg: 428, MaxDmg: 477, Mana: 265, DamageType: DamageTypeNature},
@@ -104,11 +168,14 @@ var spells = []Spell{
 	{ID: MagicIDTLCLB, Name: "TLCLB", Coeff: 0.0, CastTime: 0, MinDmg: 694, MaxDmg: 807, Mana: 0, DamageType: DamageTypeNature},
 }
 
+// Spell lookup map to make lookups faster.
 var spellmap = map[int32]*Spell{}
 
 func init() {
 	for _, sp := range spells {
-		sp2 := sp //wtf go?
+		// Turns out to increase efficiency go 'range' will actually only allocate a single struct and mutate.
+		// If we want to create a pointer we need to clone the struct.
+		sp2 := sp
 		spp := &sp2
 		spellmap[sp.ID] = spp
 	}
